Moving the DepositDataAccessInterface to the Deposit use_case to satisfy CA
