+++ stock-tracker-app/src/main/java/use_case/deposit/DepositDataAccessInterface.java
@@ -1,4 +1,4 @@
-package data_access;
+package use_case.deposit;
 
 import java.util.Date;
 import java.util.TreeMap;
