+++ stock-tracker-app/src/main/java/use_case/deposit/DepositInteractor.java
@@ -1,6 +1,5 @@
 package use_case.deposit;
 
-import data_access.DepositDataAccessInterface;
 import entity.BankingTransaction;
 import entity.UserFactory;
